bpo-40275: Avoid importing socket in test.support * Move socket related functions from test.support to socket_helper. * Import socket and nntplib lazily in transient_internet(). * Remove importing multiprocess.
